Removed unnecessary Eq bound on InstructionLivenessBuilder.
I was considering using roaring bitmaps for the sets of reigster names, but I'd need to either create a wrapper for it or index it with usize, and I'd rather not.
But for whatever reason they don't implement Eq, and I don't need to either, so I might as well not.